+++ src/include/LatencyBenchmark.h
@@ -54,7 +54,7 @@ namespace xmem {
 			size_t len,
 			uint32_t iterations,
 #ifdef USE_SIZE_BASED_BENCHMARKS
-			uint64_t passes_per_iteration,
+			uint32_t passes_per_iteration,
 #endif
 			uint32_t num_worker_threads,
 			uint32_t mem_node,
@@ -62,7 +62,7 @@ namespace xmem {
 			pattern_mode_t pattern_mode,
 			rw_mode_t rw_mode,
 			chunk_size_t chunk_size,
-			int64_t stride_size,
+			int32_t stride_size,
 			std::vector<PowerReader*> dram_power_readers,
 			std::string name
 		);
